requirements.txt
aiohttp==3.9.1
aioredis==1.3.1
aiosignal==1.3.1
alembic==1.13.0
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
astroid==3.1.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
autoflake==2.3.0
bcrypt==4.1.2
billiard==4.2.0
black==24.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
dill==0.3.8
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-cache==0.1.0
fastapi-cache2==0.2.1
flake8==7.0.0
flower==2.0.1
frozenlist==1.4.1
greenlet==3.0.2
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jose==1.0.0
jwt==1.3.1
kombu==5.3.5
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyflakes==3.2.0
PyJWT==2.8.0
pylint==3.1.0
pytest==8.0.0
pytest-asyncio==0.23.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
rsa==4.9
sentry-sdk==1.40.6
six==1.16.0
sniffio==1.3.0
sqladmin==0.16.0
SQLAlchemy==2.0.23
starlette==0.27.0
time-machine==2.13.0
tomlkit==0.12.4
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.2.1
uvicorn==0.24.0.post1
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
WTForms==3.1.2
yarl==1.9.4
